Fixing minor bugs

+++ app/src/main/java/com/bs/sriwilispetugas/data/repository/modelhelper/CardDetailPesanan.kt
@@ -2,5 +2,5 @@ package com.bs.sriwilispetugas.data.repository.modelhelper
 
 data class CardDetailPesanan(
     val nama_kategori: String,
-    val berat: Double
+    val berat: Float
 )
