+++ app/src/main/java/com/bs/sriwilispetugas/data/repository/modelhelper/CardPesanan.kt
@@ -10,5 +10,5 @@ data class CardPesanan(
     val alamat_nasabah: String,
     val nama_nasabah: String,
     val no_hp_nasabah: String,
-    val total_berat: Double
+    val total_berat: Float
 )
