+++ app/src/main/java/com/bs/sriwilispetugas/data/response/PesananSampahResponse.kt
@@ -24,7 +24,7 @@ data class PesananSampah(
     val id: Int,
     val id_pesanan_sampah_keranjang: String,
     val kategori: String,
-    val berat_perkiraan: Int,
+    val berat_perkiraan: Float,
     val harga_perkiraan: String,
     val gambar: String?,
     val created_at: String,
